+++ packages/clerk-js/src/core/clerk.ts
@@ -20,7 +20,7 @@ import type {
   __experimental_UserVerificationModalProps,
   __experimental_UserVerificationProps,
   ActiveSessionResource,
-  AuthenticateWithCoinbaseParams,
+  AuthenticateWithCoinbaseWalletParams,
   AuthenticateWithGoogleOneTapParams,
   AuthenticateWithMetamaskParams,
   Clerk as ClerkInterface,
@@ -74,7 +74,7 @@ import {
   createPageLifecycle,
   disabledOrganizationsFeature,
   errorThrower,
-  generateSignatureWithCoinbase,
+  generateSignatureWithCoinbaseWallet,
   generateSignatureWithMetamask,
   getClerkQueryParam,
   getWeb3Identifier,
@@ -331,7 +331,12 @@ export class Clerk implements ClerkInterface {
     const cb = typeof callbackOrOptions === 'function' ? callbackOrOptions : defaultCb;
 
     if (!opts.sessionId || this.client.activeSessions.length === 1) {
-      await this.client.destroy();
+      if (this.#options.experimental?.persistClient) {
+        await this.client.removeSessions();
+      } else {
+        await this.client.destroy();
+      }
+
       return this.setActive({
         session: null,
         beforeEmit: ignoreEventValue(cb),
@@ -1399,8 +1404,8 @@ export class Clerk implements ClerkInterface {
     await this.authenticateWithWeb3({ ...props, strategy: 'web3_metamask_signature' });
   };
 
-  public authenticateWithCoinbase = async (props: AuthenticateWithCoinbaseParams = {}): Promise<void> => {
-    await this.authenticateWithWeb3({ ...props, strategy: 'web3_coinbase_signature' });
+  public authenticateWithCoinbaseWallet = async (props: AuthenticateWithCoinbaseWalletParams = {}): Promise<void> => {
+    await this.authenticateWithWeb3({ ...props, strategy: 'web3_coinbase_wallet_signature' });
   };
 
   public authenticateWithWeb3 = async ({
@@ -1415,7 +1420,9 @@ export class Clerk implements ClerkInterface {
     }
     const provider = strategy.replace('web3_', '').replace('_signature', '') as Web3Provider;
     const identifier = await getWeb3Identifier({ provider });
-    const generateSignature = provider === 'metamask' ? generateSignatureWithMetamask : generateSignatureWithCoinbase;
+    const generateSignature =
+      provider === 'metamask' ? generateSignatureWithMetamask : generateSignatureWithCoinbaseWallet;
+
     const navigate = (to: string) =>
       customNavigate && typeof customNavigate === 'function' ? customNavigate(to) : this.navigate(to);