+++ packages/react/src/isomorphicClerk.ts
@@ -6,7 +6,7 @@ import type {
   __experimental_UserVerificationModalProps,
   __experimental_UserVerificationProps,
   ActiveSessionResource,
-  AuthenticateWithCoinbaseParams,
+  AuthenticateWithCoinbaseWalletParams,
   AuthenticateWithGoogleOneTapParams,
   AuthenticateWithMetamaskParams,
   Clerk,
@@ -95,7 +95,7 @@ type IsomorphicLoadedClerk = Without<
   | 'handleGoogleOneTapCallback'
   | 'handleUnauthenticated'
   | 'authenticateWithMetamask'
-  | 'authenticateWithCoinbase'
+  | 'authenticateWithCoinbaseWallet'
   | 'authenticateWithWeb3'
   | 'authenticateWithGoogleOneTap'
   | 'createOrganization'
@@ -117,7 +117,7 @@ type IsomorphicLoadedClerk = Without<
   handleUnauthenticated: () => void;
   // TODO: Align Promise unknown
   authenticateWithMetamask: (params: AuthenticateWithMetamaskParams) => Promise<void>;
-  authenticateWithCoinbase: (params: AuthenticateWithCoinbaseParams) => Promise<void>;
+  authenticateWithCoinbaseWallet: (params: AuthenticateWithCoinbaseWalletParams) => Promise<void>;
   authenticateWithWeb3: (params: ClerkAuthenticateWithWeb3Params) => Promise<void>;
   authenticateWithGoogleOneTap: (
     params: AuthenticateWithGoogleOneTapParams,
@@ -1063,12 +1063,12 @@ export class IsomorphicClerk implements IsomorphicLoadedClerk {
     }
   };
 
-  authenticateWithCoinbase = async (params: AuthenticateWithCoinbaseParams): Promise<void> => {
-    const callback = () => this.clerkjs?.authenticateWithCoinbase(params);
+  authenticateWithCoinbaseWallet = async (params: AuthenticateWithCoinbaseWalletParams): Promise<void> => {
+    const callback = () => this.clerkjs?.authenticateWithCoinbaseWallet(params);
     if (this.clerkjs && this.#loaded) {
       return callback() as Promise<void>;
     } else {
-      this.premountMethodCalls.set('authenticateWithCoinbase', callback);
+      this.premountMethodCalls.set('authenticateWithCoinbaseWallet', callback);
     }
   };