+++ packages/ui/tsup.config.ts
@@ -23,7 +23,6 @@ const tailwindcssTransformerCode = {
     build.onEnd(async () => {
       const styleSheet = await generateStylesheet(styleCache, {
         tailwindConfig: path.join(process.cwd(), 'src', 'tailwind.config.ts'),
-        globalCss: fs.readFileSync(path.join(process.cwd(), 'src', 'global.css'), 'utf8'),
       });
       await fs.promises.mkdir(outDir, { recursive: true });
       await fs.promises.writeFile(path.join(outDir, 'styles.css'), styleSheet);
@@ -40,16 +39,19 @@ export default defineConfig(overrideOptions => {
       PACKAGE_NAME: `"${name}"`,
       PACKAGE_VERSION: `"${version}"`,
       __DEV__: `${!isProd}`,
+      'import.meta.vitest': 'undefined',
     },
     dts: true,
     entry: {
       'components/sign-in': 'src/components/sign-in/index.tsx',
       'components/sign-up': 'src/components/sign-up/index.tsx',
+      contexts: 'src/contexts/index.ts',
     },
     external: ['react', 'react-dom'],
     format: ['cjs', 'esm'],
     minify: false,
     sourcemap: true,
     esbuildPlugins: [tailwindcssTransformerCode],
+    treeshake: true,
   };
 });